Downsize container instances based on all previous running revisions (#661)
Downsize now considers all previous revisions.

+++ rest/src/main/java/software/wings/service/impl/instance/InstanceHelper.java
@@ -159,7 +159,7 @@ public class InstanceHelper {
             .withInfraMappingType(infraMappingType)
             .withLastPipelineExecutionId(pipelineSummary == null ? null : pipelineSummary.getPipelineId())
             .withLastPipelineExecutionName(pipelineSummary == null ? null : pipelineSummary.getPipelineName())
-            .withLastDeployedAt(phaseExecutionData.getEndTs().longValue())
+            .withLastDeployedAt(phaseExecutionData.getEndTs())
             .withLastDeployedById(triggeredBy.getUuid())
             .withLastDeployedByName(triggeredBy.getName())
             .withServiceId(phaseExecutionData.getServiceId())
